Add the "add api key" control
