+++ src/Bootstrap/less/theme/base.less
@@ -166,7 +166,7 @@ body {
   }
 }
 
-h2 {
+h2, h3 {
   .ms-Icon {
     font-size: 0.6em;
     position: relative;
