+++ src/NuGetGallery/ViewModels/ApiKeysViewModel.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
 using System.Collections.Generic;
 
 namespace NuGetGallery
@@ -10,21 +9,7 @@ namespace NuGetGallery
     {
         public string SiteRoot { get; set; }
         public IList<ApiKeyViewModel> ApiKeys { get; set; }
-    }
-
-    public class ApiKeyViewModel
-    {
-        public bool IsNonScopedV1ApiKey { get; set; }
-        public bool HasExpired { get; set; }
-        public string Expires { get; set; }
-        public IList<string> ShortScopeList { get; set; }
-        public IList<string> Scopes { get; set; }
-        public IList<string> ShortPackageList { get; set; }
-        public IList<IList<string>> PackageBatches { get; set; }
-        public string GlobPattern { get; set; }
-        public string Value { get; set; }
-        public string Type { get; set; }
-        public int Key { get; set; }
-        public string Description { get; set; }
+        public int ExpirationInDaysForApiKeyV1 { get; set; }
+        public IList<string> PackageIds { get; set; }
     }
 }
